Refactore models
