+++ src/openprocurement/api/models/auction_models/models.py
@@ -1,25 +1,37 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime
-from decimal import Decimal, InvalidOperation, ROUND_HALF_UP
-from hashlib import algorithms, new as hash_new
 from string import hexdigits
 from urlparse import urlparse, parse_qs
 from uuid import uuid4
 
 from couchdb_schematics.document import SchematicsDocument
-from iso8601 import parse_date, ParseError
 from schematics.exceptions import ConversionError, ValidationError
-from schematics.models import Model as SchematicsModel
-from schematics.transforms import whitelist, blacklist, export_loop, convert
-from schematics.types import (StringType, FloatType, URLType, IntType,
-                              BooleanType, BaseType, EmailType, MD5Type, DecimalType as BaseDecimalType)
-from schematics.types.compound import (ModelType, DictType,
-                                       ListType as BaseListType)
+from schematics.transforms import whitelist, blacklist, export_loop
+from schematics.types import (
+    StringType,
+    FloatType,
+    URLType,
+    IntType,
+    BooleanType,
+    BaseType,
+    EmailType,
+    MD5Type
+)
+from schematics.types.compound import (
+    ModelType,
+    DictType
+)
 from schematics.types.serializable import serializable
 
 from openprocurement.api.constants import (
     CPV_CODES, ORA_CODES, TZ, DK_CODES, CPV_BLOCK_FROM, ATC_CODES, INN_CODES, ATC_INN_CLASSIFICATIONS_FROM,
 )
+from openprocurement.api.models.schematics_extender import (
+    Model,
+    ListType,
+    IsoDateTimeType,
+    HashType
+)
+from openprocurement.api.models.models import Period
 from openprocurement.api.utils import get_now, set_parent, get_schematics_document
 
 schematics_default_role = SchematicsDocument.Options.roles['default'] + blacklist("__parent__")
@@ -30,183 +42,6 @@ listing_role = whitelist('dateModified', 'doc_id')
 draft_role = whitelist('status')
 
 
-class DecimalType(BaseDecimalType):
-    def __init__(self, precision=-3, min_value=None, max_value=None, **kwargs):
-        self.min_value, self.max_value = min_value, max_value
-        self.precision = Decimal("1E{:d}".format(precision))
-        super(DecimalType, self).__init__(**kwargs)
-
-    def to_primitive(self, value, context=None):
-        return value
-
-    def to_native(self, value, context=None):
-        try:
-            value = Decimal(value).quantize(self.precision, rounding=ROUND_HALF_UP).normalize()
-        except (TypeError, InvalidOperation):
-            raise ConversionError(self.messages['number_coerce'].format(value))
-
-        if self.min_value is not None and value < self.min_value:
-            raise ConversionError(self.messages['number_min'].format(value))
-        if self.max_value is not None and self.max_value < value:
-            raise ConversionError(self.messages['number_max'].format(value))
-
-        return value
-
-
-class IsoDateTimeType(BaseType):
-    MESSAGES = {
-        'parse': u'Could not parse {0}. Should be ISO8601.',
-    }
-
-    def to_native(self, value, context=None):
-        if isinstance(value, datetime):
-            return value
-        try:
-            date = parse_date(value, None)
-            if not date.tzinfo:
-                date = TZ.localize(date)
-            return date
-        except ParseError:
-            raise ConversionError(self.messages['parse'].format(value))
-        except OverflowError as e:
-            raise ConversionError(e.message)
-
-    def to_primitive(self, value, context=None):
-        return value.isoformat()
-
-
-class ListType(BaseListType):
-
-    def export_loop(self, list_instance, field_converter,
-                    role=None, print_none=False):
-        """Loops over each item in the model and applies either the field
-        transform or the multitype transform.  Essentially functions the same
-        as `transforms.export_loop`.
-        """
-        data = []
-        for value in list_instance:
-            if hasattr(self.field, 'export_loop'):
-                shaped = self.field.export_loop(value, field_converter,
-                                                role=role,
-                                                print_none=print_none)
-                feels_empty = shaped and len(shaped) == 0
-            else:
-                shaped = field_converter(self.field, value)
-                feels_empty = shaped is None
-
-            # Print if we want empty or found a value
-            if feels_empty and self.field.allow_none():
-                data.append(shaped)
-            elif shaped is not None:
-                data.append(shaped)
-            elif print_none:
-                data.append(shaped)
-
-        # Return data if the list contains anything
-        if len(data) > 0:
-            return data
-        elif len(data) == 0 and self.allow_none():
-            return data
-        elif print_none:
-            return data
-
-
-class SifterListType(ListType):
-    def __init__(self, field, min_size=None, max_size=None,
-                 filter_by=None, filter_in_values=[], **kwargs):
-        self.filter_by = filter_by
-        self.filter_in_values = filter_in_values
-        super(SifterListType, self).__init__(field, min_size=min_size,
-                                             max_size=max_size, **kwargs)
-
-    def export_loop(self, list_instance, field_converter,
-                    role=None, print_none=False):
-        """ Use the same functionality as original method but apply
-        additional filters.
-        """
-        data = []
-        for value in list_instance:
-            if hasattr(self.field, 'export_loop'):
-                item_role = role
-                # apply filters
-                if role not in ['plain', None] and self.filter_by and hasattr(value, self.filter_by):
-                    val = getattr(value, self.filter_by)
-                    if val in self.filter_in_values:
-                        item_role = val
-
-                shaped = self.field.export_loop(value, field_converter,
-                                                role=item_role,
-                                                print_none=print_none)
-                feels_empty = shaped and len(shaped) == 0
-            else:
-                shaped = field_converter(self.field, value)
-                feels_empty = shaped is None
-
-            # Print if we want empty or found a value
-            if feels_empty and self.field.allow_none():
-                data.append(shaped)
-            elif shaped is not None:
-                data.append(shaped)
-            elif print_none:
-                data.append(shaped)
-
-        # Return data if the list contains anything
-        if len(data) > 0:
-            return data
-        elif len(data) == 0 and self.allow_none():
-            return data
-        elif print_none:
-            return data
-
-
-class Model(SchematicsModel):
-    class Options(object):
-        """Export options for Document."""
-        serialize_when_none = False
-        roles = {
-            "default": blacklist("__parent__"),
-            "embedded": blacklist("__parent__"),
-        }
-
-    __parent__ = BaseType()
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            for k in self._fields:
-                if k != '__parent__' and self.get(k) != other.get(k):
-                    return False
-            return True
-        return NotImplemented
-
-    def convert(self, raw_data, **kw):
-        """
-        Converts the raw data into richer Python constructs according to the
-        fields on the model
-        """
-        value = convert(self.__class__, raw_data, **kw)
-        for i, j in value.items():
-            if isinstance(j, list):
-                for x in j:
-                    set_parent(x, self)
-            else:
-                set_parent(j, self)
-        return value
-
-    def to_patch(self, role=None):
-        """
-        Return data as it would be validated. No filtering of output unless
-        role is defined.
-        """
-        field_converter = lambda field, value: field.to_primitive(value)
-        data = export_loop(self.__class__, self, field_converter, role=role, raise_error_on_role=True, print_none=True)
-        return data
-
-    def get_role(self):
-        root = self.__parent__
-        while root.__parent__ is not None:
-            root = root.__parent__
-        request = root.request
-        return 'Administrator' if request.authenticated_role == 'Administrator' else 'edit'
 
 
 class Value(Model):
@@ -252,26 +87,6 @@ class Feature(Model):
             raise ValidationError(u"relatedItem should be one of items")
         if data.get('featureOf') == 'lot' and isinstance(data['__parent__'], Model) and relatedItem not in [i.id for i in data['__parent__'].lots]:
             raise ValidationError(u"relatedItem should be one of lots")
-
-
-class Guarantee(Model):
-    amount = FloatType(required=True, min_value=0)  # Amount as a number.
-    currency = StringType(required=True, default=u'UAH', max_length=3, min_length=3)  # The currency in 3-letter ISO 4217 format.
-
-
-class Period(Model):
-    """The period when the tender is open for submissions. The end date is the closing date for tender submissions."""
-
-    startDate = IsoDateTimeType()  # The state date for the period.
-    endDate = IsoDateTimeType()  # The end date for the period.
-
-    def validate_startDate(self, data, value):
-        if value and data.get('endDate') and data.get('endDate') < value:
-            raise ValidationError(u"period should begin before its end")
-
-
-class PeriodEndRequired(Period):
-    endDate = IsoDateTimeType(required=True)  # The end date for the period.
 
 
 class Classification(Model):
@@ -364,34 +179,6 @@ class Location(Model):
     latitude = BaseType(required=True)
     longitude = BaseType(required=True)
     elevation = BaseType()
-
-
-class HashType(StringType):
-
-    MESSAGES = {
-        'hash_invalid': "Hash type is not supported.",
-        'hash_length': "Hash value is wrong length.",
-        'hash_hex': "Hash value is not hexadecimal.",
-    }
-
-    def to_native(self, value, context=None):
-        value = super(HashType, self).to_native(value, context)
-
-        if ':' not in value:
-            raise ValidationError(self.messages['hash_invalid'])
-
-        hash_type, hash_value = value.split(':', 1)
-
-        if hash_type not in algorithms:
-            raise ValidationError(self.messages['hash_invalid'])
-
-        if len(hash_value) != hash_new(hash_type).digest_size * 2:
-            raise ValidationError(self.messages['hash_length'])
-        try:
-            int(hash_value, 16)
-        except ValueError:
-            raise ConversionError(self.messages['hash_hex'])
-        return value
 
 
 class Document(Model):
